chore(migration): signup.spec to ts
